requirements.txt
albucore==0.0.13
albumentations==1.4.12
altair==5.3.0
annotated-types==0.7.0
attrs==24.1.0
blinker==1.8.2
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
eval_type_backport==0.2.0
filelock==3.13.1
fonttools==4.53.1
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.43
idna==3.7
imageio==2.34.2
Jinja2==3.1.3
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
mdurl==0.1.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.1
pandas==2.2.2
pillow==10.2.0
protobuf==5.27.3
pyarrow==17.0.0
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.19.1
scikit-image==0.24.0
scipy==1.14.0
six==1.16.0
smmap==5.0.1
streamlit==1.37.0
sympy==1.12
tenacity==8.5.0
tifffile==2024.7.24
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
torch==2.4.0+cpu
torchaudio==2.4.0+cpu
torchvision==0.19.0+cpu
tornado==6.4.1
tqdm==4.66.5
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.2
watchdog==4.0.1
